// src/common/types/types.ts
export enum RoleTypes {
    user = 'user',
    admin = 'admin',
}
export enum GenderTypes {
    male = 'male',
    female = 'female',
}

export enum OTPTypes {
    confirmation = 'confirmation',
    resetPassword = 'resetPassword',
}

export enum PaymentMethodsTypes {
    cash = 'cash',
    card = 'card',
}


export enum OrderStatusTypes {
    pending = 'pending',
    shipped = 'shipped',
    delivered = 'delivered',
    onWay = 'onWay',
    cancelled = 'cancelled',
    rejected = 'rejected',
    refunded = 'refunded',
    paid = 'paid',
    placed = 'placed',
    cash = 'cash',
}